--- redux/action.go
@@ -1,6 +1,0 @@
-package redux
-
-/*
-Action is a value that is dispatched to the redux store
-*/
-type Action interface{}
