--- redux/dispatch.go
@@ -1,9 +1,0 @@
-package redux
-
-/*
-DispatchHandler will handle every dispatch
-*/
-type DispatchHandler func(
-	store *Store,
-	action Action,
-)
