--- redux/state.go
@@ -1,6 +1,0 @@
-package redux
-
-/*
-State is the state of a redux store
-*/
-type State interface{}
